move from script to init to keep imports simple for users
